Fix statement for 64 bit platforms where size_t != unsigned int.
git-svn-id: https://svn.dealii.org/trunk@4781 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/lac/source/compressed_sparsity_pattern.cc
@@ -104,7 +104,7 @@ CompressedSparsityPattern::max_entries_per_row () const
 {
   unsigned int m = 0;
   for (unsigned int i=1; i<rows; ++i)
-    m = std::max (m, column_indices[i].size());
+    m = std::max (m, static_cast<unsigned int>(column_indices[i].size()));
 
   return m;
 };
